Phase1Load/requirements.txt
greenlet==3.1.1
numpy==2.2.3
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
